Publish per-port latency observations
