+++ src/rust/shoop_engine/src/lib.rs
@@ -42,6 +42,7 @@ pub mod graph;
 pub mod graph_build;
 pub mod graph_scheduler;
 pub mod internal_audio_port;
+pub mod latency_runtime;
 pub mod loop_mode;
 pub mod midi;
 pub mod midi_buffering_input_port;
@@ -127,6 +128,7 @@ pub use fx_chain::{FXChainState, FXChainType};
 pub use graph::{processing_order, GraphError, NodeIdx, NodeSpec};
 pub use graph_build::{ChannelDesc, GraphDesc, LoopDesc, PortDesc};
 pub use internal_audio_port::InternalAudioPort;
+pub use latency_runtime::{AtomicLatencyObservation, RuntimeLatencyObservation};
 pub use loop_mode::LoopMode;
 pub use midi_buffering_input_port::MidiBufferingInputPort;
 #[cfg(any(feature = "app_backend", feature = "native_audio_backend"))]
